* First round of listApp working * Change Job id type from int32 to long, even in database!   => Schema migration required
* This is awesome example
  https://github.com/saturnism/grpc-java-demos

TODOs
* Add access logging on the server side
* Add switch on client and server side
* Add more APIs
* Enable authentication
* Enable default/config cert files for gRPC2
* Other authentication method like OAuth2?

+++ retz-client/src/main/java/io/github/retz/cli/CommandKillall.java
@@ -100,11 +100,11 @@ public class CommandKillall implements SubCommand {
             LOG.info("{}: {} jobs killed, {} jobs failed", state, killed.size(), failed.size());
 
             if (verbose) {
-                String k = killed.stream().map(job -> Integer.toString(job.id())).collect(Collectors.joining(", "));
+                String k = killed.stream().map(job -> Long.toString(job.id())).collect(Collectors.joining(", "));
                 LOG.info("Jobs killed: {}", k);
             }
             if (!failed.isEmpty()) {
-                String f = failed.stream().map(job -> Integer.toString(job.id())).collect(Collectors.joining(", "));
+                String f = failed.stream().map(job -> Long.toString(job.id())).collect(Collectors.joining(", "));
                 LOG.error("Failed to kill jobs: [{}]", f);
             }
